refactor/EV-1077/moving types into protocols file
