--- src/utils/types.ts
@@ -1,6 +1,0 @@
-import { IRoutesController } from "@/presentation/routes/Controller";
-import { Request, Response } from "express";
-
-export type ControllerAdapterType = (
-  controller: IRoutesController
-) => (req: Request, res: Response) => Response;
